requirements.txt
livekit-agents>=0.12.16
livekit-plugins-openai>=0.10.17
livekit-plugins-cartesia>=0.4.7
livekit-plugins-deepgram>=0.6.17
livekit-plugins-silero>=0.7.4
livekit-plugins-turn-detector>=0.4.0
livekit-plugins-elevenlabs>=0.7.1
python-dotenv~=1.0
sib-api-v3-sdk>=7.6.0